+++ src/interfaces/actor-data.ts
@@ -123,7 +123,10 @@ export interface CharacterCommon {
     hasImpLevelHeaded: boolean;
     hasQuick: boolean;
   };
-  powerPoints: any; //FIXME Find better type
+  powerPoints: {
+    value: number;
+    max: number;
+  } & Record<string, { value: number; max: number }>;
   additionalStats: Partial<Record<string, AdditionalStat>>;
   wildcard: boolean;
 }