+++ src/module/Benny.ts
@@ -1,5 +1,5 @@
 export default class Benny extends DiceTerm {
-  constructor(termData) {
+  constructor(termData: DiceTerm.TermData) {
     termData.faces = 2;
     super(termData);
   }