+++ src/module/dialog/entity-tweaks.ts
@@ -93,8 +93,7 @@ export default class SwadeEntityTweaks extends FormApplication {
     );
 
     // Update the actor
-    //FIXME revisit later
-    //@ts-ignore
+    //@ts-expect-error This can update both Item and Actor Documents but I'm not sure how to make TS understand
     await this.object.update(expandedFormData);
     this.object.sheet.render(true);
   }
@@ -114,7 +113,7 @@ export default class SwadeEntityTweaks extends FormApplication {
     const formFields =
       getProperty(expandedFormData, 'data.additionalStats') || {};
     const prototypeFields = this._getAppropriateSettingFields();
-    const newFields = duplicate(
+    const newFields = deepClone(
       getProperty(this.object.data, 'data.additionalStats'),
     );
     //handle setting specific fields