+++ src/module/sheets/SwadeItemSheet.ts
@@ -12,11 +12,10 @@ export default class SwadeItemSheet extends ItemSheet {
   }
 
   static get defaultOptions() {
-    //TODO Revisit once mergeObject is typed correctly
-    //@ts-ignore
-    return mergeObject(super.defaultOptions, {
+    return {
+      ...super.defaultOptions,
       width: 560,
-      height: 'auto',
+      height: 'auto' as 'auto',
       classes: ['swade', 'sheet', 'item'],
       tabs: [
         {
@@ -27,7 +26,7 @@ export default class SwadeItemSheet extends ItemSheet {
       ],
       scrollY: ['.actions-list'],
       resizable: true,
-    });
+    };
   }
 
   /**
@@ -167,11 +166,14 @@ export default class SwadeItemSheet extends ItemSheet {
       if (!modifier.match(/^[+-]/)) {
         modifier = '+' + modifier;
       }
-      const dieSides = statData.value || 4;
-      new Roll(`1d${dieSides}${modifier}`).roll().toMessage({
-        speaker: ChatMessage.getSpeaker(),
-        flavor: `${this.item.name} - ${statData.label}`,
-      });
+      //return of there's no value to roll
+      if (!statData.value) return;
+      new Roll(`1d${statData.value}${modifier}`)
+        .evaluate({ async: false })
+        .toMessage({
+          speaker: ChatMessage.getSpeaker(),
+          flavor: `${this.item.name} - ${statData.label}`,
+        });
     });
   }
 
@@ -188,7 +190,7 @@ export default class SwadeItemSheet extends ItemSheet {
     if (ownerIsWildcard || !this.item.isOwned) {
       data.data.ownerIsWildcard = true;
     }
-    const additionalStats = data.data.additionalStats || {};
+    const additionalStats = data.data.data.additionalStats || {};
     for (const attr of Object.values(additionalStats)) {
       attr['isCheckbox'] = attr['dtype'] === 'Boolean';
     }
@@ -232,7 +234,7 @@ export default class SwadeItemSheet extends ItemSheet {
         const pack = game.packs.get(data.pack) as Compendium;
         item = (await pack.getEntity(data.id)) as SwadeItem;
       } else if ('actorId' in data) {
-        item = new SwadeItem(data.data, {});
+        item = new SwadeItem(data.data);
       } else {
         item = game.items.get(data.id) as SwadeItem;
       }
@@ -251,7 +253,8 @@ export default class SwadeItemSheet extends ItemSheet {
     }
 
     //prep item data
-    const itemData = duplicate(item.data);
+    //@ts-ignore
+    const itemData = deepClone(item.data.toObject());
     delete itemData['_id'];
     delete itemData['permission'];